Se crea nueva clase PrezGUI y se extrae de Prez la parte correspondiente a tal clase. Se ha generado un ejecutable y parece funcionar bien.
